+++ ports/stm32/i2c.c
@@ -590,4 +590,72 @@ int i2c_find_peripheral(mp_obj_t id) {
     return i2c_id;
 }
 
+#if MICROPY_PY_PYB_LEGACY
+
+#if defined(MICROPY_HW_I2C1_SCL)
+void I2C1_EV_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C1_EV_IRQn > 0);
+    IRQ_ENTER(I2C1_EV_IRQn);
+    i2c_ev_irq_handler(1);
+    IRQ_EXIT(I2C1_EV_IRQn);
+}
+
+void I2C1_ER_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C1_ER_IRQn > 0);
+    IRQ_ENTER(I2C1_ER_IRQn);
+    i2c_er_irq_handler(1);
+    IRQ_EXIT(I2C1_ER_IRQn);
+}
+#endif // defined(MICROPY_HW_I2C1_SCL)
+
+#if defined(MICROPY_HW_I2C2_SCL)
+void I2C2_EV_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C2_EV_IRQn > 0);
+    IRQ_ENTER(I2C2_EV_IRQn);
+    i2c_ev_irq_handler(2);
+    IRQ_EXIT(I2C2_EV_IRQn);
+}
+
+void I2C2_ER_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C2_ER_IRQn > 0);
+    IRQ_ENTER(I2C2_ER_IRQn);
+    i2c_er_irq_handler(2);
+    IRQ_EXIT(I2C2_ER_IRQn);
+}
+#endif // defined(MICROPY_HW_I2C2_SCL)
+
+#if defined(MICROPY_HW_I2C3_SCL)
+void I2C3_EV_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C3_EV_IRQn > 0);
+    IRQ_ENTER(I2C3_EV_IRQn);
+    i2c_ev_irq_handler(3);
+    IRQ_EXIT(I2C3_EV_IRQn);
+}
+
+void I2C3_ER_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C3_ER_IRQn > 0);
+    IRQ_ENTER(I2C3_ER_IRQn);
+    i2c_er_irq_handler(3);
+    IRQ_EXIT(I2C3_ER_IRQn);
+}
+#endif // defined(MICROPY_HW_I2C3_SCL)
+
+#if defined(MICROPY_HW_I2C4_SCL)
+void I2C4_EV_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C4_EV_IRQn > 0);
+    IRQ_ENTER(I2C4_EV_IRQn);
+    i2c_ev_irq_handler(4);
+    IRQ_EXIT(I2C4_EV_IRQn);
+}
+
+void I2C4_ER_IRQHandler(void) {
+    MP_STATIC_ASSERT(I2C4_ER_IRQn > 0);
+    IRQ_ENTER(I2C4_ER_IRQn);
+    i2c_er_irq_handler(4);
+    IRQ_EXIT(I2C4_ER_IRQn);
+}
+#endif // defined(MICROPY_HW_I2C4_SCL)
+
+#endif // MICROPY_PY_PYB_LEGACY
+
 #endif // MICROPY_HW_ENABLE_HW_I2C
